Also fix compilation for ogl-es 2.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/ogl-es@4501 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COGLES2MaterialRenderer.cpp
@@ -397,7 +397,7 @@ bool COGLES2MaterialRenderer::setPixelShaderConstant(s32 index, const f32* float
 			{
 				if(floats)
 				{
-					const GLint id = *floats;
+					const GLint id = (GLint)(*floats);
 					glUniform1iv(UniformInfo[index].location, 1, &id);
 				}
 				else
